jsonapi_demo_cli/requirements.txt
oauthlib==3.0.0
jsonapi-requests==0.6.0
